# lib/api-zod/src/generated/types/cronResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Third Place — Coffee Shop OS API
 * OpenAPI spec version: 0.1.0
 */

export interface CronResult {
  success: boolean;
  message: string;
  found?: number | null;
  validated?: number | null;
  updated?: number | null;
}