# lib/api-zod/src/generated/types/discoveryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Third Place — Coffee Shop OS API
 * OpenAPI spec version: 0.1.0
 */
import type { DiscoveryItemStatus } from "./discoveryItemStatus";

export interface DiscoveryItem {
  id: number;
  name: string;
  address?: string | null;
  neighborhood?: string | null;
  sourceUrl?: string | null;
  discoveredAt: string;
  status: DiscoveryItemStatus;
}