# lib/api-zod/src/generated/types/venue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Third Place — Coffee Shop OS API
 * OpenAPI spec version: 0.1.0
 */
import type { VenueCategory } from "./venueCategory";
import type { VenueStatus } from "./venueStatus";

export interface Venue {
  id: number;
  name: string;
  neighborhood?: string | null;
  address?: string | null;
  category: VenueCategory;
  tags: string[];
  workRating?: number | null;
  coffeeRating?: number | null;
  hours?: string | null;
  mapsUrl?: string | null;
  sourceUrl?: string | null;
  status: VenueStatus;
  dateAdded?: string | null;
  visitCount: number;
  lastVisited?: string | null;
  notes?: string | null;
}