# lib/api-zod/src/generated/types/venueStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Third Place — Coffee Shop OS API
 * OpenAPI spec version: 0.1.0
 */
import type { Venue } from "./venue";
import type { VenueStatsByCategory } from "./venueStatsByCategory";

export interface VenueStats {
  total: number;
  active: number;
  closed: number;
  unverified: number;
  byCategory: VenueStatsByCategory;
  recentlyAdded: Venue[];
}