# lib/api-zod/src/generated/types/visit.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Third Place — Coffee Shop OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Visit {
  id: number;
  venueId: number;
  visitedAt: string;
  workedOn?: string | null;
  accomplished?: string | null;
  whoSaw?: string | null;
  coffeeNotes?: string | null;
}