# lib/api-zod/src/generated/types/visitStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Third Place — Coffee Shop OS API
 * OpenAPI spec version: 0.1.0
 */

export interface VisitStats {
  totalVisits: number;
  uniqueVenues: number;
}