# lib/api-zod/src/generated/types/visitWithVenue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Third Place — Coffee Shop OS API
 * OpenAPI spec version: 0.1.0
 */
import type { Venue } from "./venue";
import type { Visit } from "./visit";

export type VisitWithVenue = Visit & {
  venue: Venue;
};